making the iterative solver deriving from a Laplacian functor
Stuff to think about:
- give the Iterativesolve a raw functor instead of a raw pointer
- move back grid redefinition to self gravity
