+++ src/gravity/selfGravity.hpp
@@ -15,6 +15,7 @@
 #include "grid.hpp"
 #include "fluid_defs.hpp"
 #include "iterativesolver.hpp"
+#include "laplacian.hpp"
 
 #ifdef WITH_MPI
 #include "mpi.hpp"
@@ -25,52 +26,22 @@ class DataBlock;
 
 class SelfGravity {
  public:
-  // Types of boundary which can be treated
-  enum GravityBoundaryType {internalgrav, periodic, nullgrad, nullpot, userdef, axis, origin};
-  enum GravitySolver {JACOBI, BICGSTAB, PBICGSTAB, PCG, CG, PMINRES, MINRES};
 
-  // Handling userdef boundary.
-  using UserDefBoundaryFunc = void (*) (DataBlock &, int dir, BoundarySide side,
-                                       const real t, IdefixArray3D<real> &arr);
+  enum GravitySolver {JACOBI, BICGSTAB, PBICGSTAB, PCG, CG, PMINRES, MINRES};
 
   SelfGravity();  // Default (empty) constructor
   void Init(Input &, DataBlock *);  // Initialisation of the class attributes
   void ShowConfig();                // display current configuration
   void InitSolver(); // (Re)initialisation of the solver for a given density distribution
 
-  void InitInternalGrid(); // initialise the extra internal grid (for origin BCs)
-  real ComputeJacobiCFL();  // Compute step for Jacobi following CFL conditions
   void SubstractMeanDensity();  // Compute and substract the average input density
-
-  void ComputeLaplacian(IdefixArray3D<real> array,
-                        IdefixArray3D<real> laplacian);
-  // Compute the laplacian of the given array
-
-  void EnforceBoundary(int dir, BoundarySide side, GravityBoundaryType type,
-                       IdefixArray3D<real> &);
-  // Enforce boundary conditions in a specific dir, side, following boundary type
-  // and for the specified array
-
-  void SetBoundaries(IdefixArray3D<real> &);  // Set the proper boundaries for the given array
-
 
   void SolvePoisson(); // Solve Poisson equation
   void AddSelfGravityPotential(IdefixArray3D<real> &);
-  void EnrollUserDefBoundary(UserDefBoundaryFunc);  // Enroll user-defined boundary conditions
-  void InitPreconditionner();   // For preconditionning versions
-  void PreComputeLaplacian();   // For faster Laplacian computation
 
-  // real gravCst; // 4*pi*G
+  IterativeSolver<Laplacian> *iterativeSolver;
 
-  // User defined Boundary conditions
-  UserDefBoundaryFunc userDefBoundaryFunc{NULL};
-  bool haveUserDefBoundary{false};
-
-  IterativeSolver<SelfGravity> *iterativeSolver;
-
-  #ifdef WITH_MPI
-  Mpi mpi;  // Mpi object when WITH_MPI is set
-  #endif
+  std::unique_ptr<Laplacian> laplacian;
 
   #ifdef DEBUG_GRAVITY
   // Used to get fields usefull for debugging
@@ -93,51 +64,14 @@ class SelfGravity {
   IdefixArray3D<real> potential;  // Gravitational potential
   IdefixArray3D<real> density;  // Density
   real dt;  // CFL timestep
-
+  
   // Local potential array size
   std::vector<int> np_tot;
-  std::vector<int> np_int;
 
-  std::vector<int> nghost;
-  std::vector<int> beg;
-  std::vector<int> end;
-
-  // offset in the left and right direction between selfgravity grid and datablock grid
-  std::vector<int> loffset;
-  std::vector<int> roffset;
-
-  // Grid for self-gravity solver (note that this grid may extend the grid of the current datablock)
-  std::vector<IdefixArray1D<real>> x;    ///< geometrical central points
-  std::vector<IdefixArray1D<real>> dx;   ///< cell width
-
-  IdefixArray1D<real> sinx2;            ///< sinx2 (only in spherical)
-
-  IdefixArray3D<real> dV;                ///< cell volume
-  std::vector<IdefixArray3D<real>> A;    ///< cell left interface area
-
-  #ifdef WITH_MPI
-  MPI_Comm originComm;                  ///< MPI communicator used by the origin boundary condition
-  #endif
-
-  bool isPeriodic; // Periodicity status of the density distribution
-  GravityBoundaryType lbound[3];  // Boundary condition to the left
-  GravityBoundaryType rbound[3];  // Boundary condition to the right
-                           // Warning : might differ from (M)HD solver !
-
-  IdefixArray3D<real> precond; // Diagonal preconditionner
-  IdefixArray4D<real> Lx1; //< Laplacian operator in x1
-  IdefixArray4D<real> Lx2; //< Laplacian operator in x2
-  IdefixArray4D<real> Lx3; //< Laplacian operator in x3
-
-  #ifdef WITH_MPI
-  IdefixArray4D<real> arr4D; // Intermediate array for boundary handling
-  #endif
-
-  bool isTwoPi{false};
+  std::array<Laplacian::LaplacianBoundaryType,3> lbound;  // Boundary condition to the left
+  std::array<Laplacian::LaplacianBoundaryType,3> rbound;  // Boundary condition to the right
 
   GravitySolver solver; // The solver  used to solve Poisson
-
-  bool havePreconditioner{false}; // Use of preconditionner (or not)
 };
 
 #endif // GRAVITY_SELFGRAVITY_HPP_
